Try to cancel subrequests; also minor cosmetics

+++ yt/ytlib/query_client/coordinator.h
@@ -33,7 +33,7 @@ TQueryStatistics CoordinateAndExecute(
     ISchemafulWriterPtr writer,
     bool isOrdered,
     std::function<std::vector<TKeyRange>(const TDataSplits&)> splitAndRegroup,
-    std::function<TEvaluateResult(const TConstQueryPtr&, size_t)> evaluateSubquery,
+    std::function<TEvaluateResult(const TConstQueryPtr&, int)> evaluateSubquery,
     std::function<TQueryStatistics(const TConstQueryPtr&, ISchemafulReaderPtr, ISchemafulWriterPtr)> evaluateTop,
     bool pushdownGroupOp = true);
